Ajout du pseudo de l'user dans l'header && ajout de la pagination (reste à faire : le bouton prev/next && securité dans  p

+++ message_sup.php
@@ -1,12 +1,13 @@
 <?php
 include('includes/connexion.inc.php');
-	
-if(isset($_GET['id']) && !empty($_GET['id'])){
+if($connect){
+	if(isset($_GET['id']) && !empty($_GET['id'])){
 
-	$query = "DELETE FROM messages WHERE id = ?";
-	$prep = $pdo->prepare($query);
-	$prep->bindValue(1, $_GET['id']);
-	$prep->execute();
+		$query = "DELETE FROM messages WHERE id = ?";
+		$prep = $pdo->prepare($query);
+		$prep->bindValue(1, $_GET['id']);
+		$prep->execute();
+	}
 }
 
 header("Location: index.php");
